// tests/protoc-gen-ts-grpc-gateway-cli/integration-everything-simple/caseProtobufES/proto/sub/message_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
// @generated from file proto/sub/message.proto (package proto.sub, syntax proto2)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file proto/sub/message.proto.
 */
export const file_proto_sub_message: GenFile =
  /*@__PURE__*/
  fileDesc(
    "Chdwcm90by9zdWIvbWVzc2FnZS5wcm90bxIJcHJvdG8uc3ViIh4KDVN0cmluZ01lc3NhZ2USDQoFdmFsdWUYASACKAlCqQEKDWNvbS5wcm90by5zdWJCDE1lc3NhZ2VQcm90b1ABWkVnaXRodWIuY29tL2dycGMtZWNvc3lzdGVtL2dycGMtZ2F0ZXdheS92Mi9leGFtcGxlcy9pbnRlcm5hbC9wcm90by9zdWKiAgNQU1iqAglQcm90by5TdWLKAglQcm90b1xTdWLiAhVQcm90b1xTdWJcR1BCTWV0YWRhdGHqAgpQcm90bzo6U3Vi",
  );

/**
 * @generated from message proto.sub.StringMessage
 */
export type StringMessage = Message<"proto.sub.StringMessage"> & {
  /**
   * @generated from field: required string value = 1;
   */
  value: string;
};

/**
 * Describes the message proto.sub.StringMessage.
 * Use `create(StringMessageSchema)` to create a new message.
 */
export const StringMessageSchema: GenMessage<StringMessage> =
  /*@__PURE__*/
  messageDesc(file_proto_sub_message, 0);